Move code into component
