+++ components/form-builder/app/FormOwnership.tsx
@@ -1,0 +1,142 @@
+import React, { ReactElement, useId, useState } from "react";
+import { useTranslation } from "next-i18next";
+import { Alert } from "@components/forms";
+import { ErrorStatus } from "@components/forms/Alert/Alert";
+import { logMessage } from "@lib/logger";
+import { Button } from "@components/globals";
+import Select from "react-select";
+import makeAnimated from "react-select/animated";
+import axios from "axios";
+import { FormRecord } from "@lib/types";
+
+interface User {
+  id: string;
+  name: string | null;
+  email: string | null;
+}
+
+interface AssignUsersToTemplateProps {
+  formRecord: FormRecord;
+  usersAssignedToFormRecord: User[];
+  allUsers: User[];
+}
+
+const usersToOptions = (users: User[]): { value: string; label: string | null }[] => {
+  return users.map((user) => {
+    return { value: user.id, label: user.email };
+  });
+};
+
+const updateUsersToTemplateAssignations = async (
+  formID: string,
+  users: { id: string; action: "add" | "remove" }[]
+) => {
+  try {
+    return await axios({
+      url: `/api/templates/${formID}`,
+      method: "PUT",
+      data: {
+        users,
+      },
+      timeout: process.env.NODE_ENV === "production" ? 60000 : 0,
+    });
+  } catch (e) {
+    logMessage.error(e);
+  }
+};
+
+export const FormOwnership = ({
+  formRecord,
+  usersAssignedToFormRecord,
+  allUsers,
+}: AssignUsersToTemplateProps) => {
+  const { t } = useTranslation("form-builder");
+
+  const [message, setMessage] = useState<ReactElement | null>(null);
+  const assignedToFormRecord = usersToOptions(usersAssignedToFormRecord);
+  const [selectedUsers, setSelectedUsers] =
+    useState<{ value: string; label: string | null }[]>(assignedToFormRecord);
+
+  const saveAssignations = async () => {
+    setMessage(null);
+    const usersToAdd: { id: string; action: "add" | "remove" }[] = selectedUsers
+      .filter((user) => {
+        return !usersAssignedToFormRecord.map((u) => u.id).includes(user.value);
+      })
+      .map((user) => {
+        return { id: user.value, action: "add" };
+      });
+
+    const usersToRemove: { id: string; action: "add" | "remove" }[] = usersAssignedToFormRecord
+      .filter((user) => {
+        return !selectedUsers.map((u) => u.value).includes(user.id);
+      })
+      .map((user) => {
+        return { id: user.id, action: "remove" };
+      });
+
+    const response = await updateUsersToTemplateAssignations(
+      formRecord.id,
+      usersToAdd.concat(usersToRemove)
+    );
+
+    if (response && response.status === 200) {
+      setMessage(
+        <Alert
+          type={ErrorStatus.SUCCESS}
+          focussable={true}
+          heading={t("responseSuccess.title")}
+          tabIndex={0}
+          className="mb-2"
+        >
+          {t("responseSuccess.message")}
+        </Alert>
+      );
+      return response.data;
+    }
+
+    setMessage(
+      <Alert
+        type={ErrorStatus.ERROR}
+        focussable={true}
+        heading={t("responseFail.title")}
+        tabIndex={0}
+        className="mb-2"
+      >
+        {t("responseFail.message")}
+      </Alert>
+    );
+  };
+
+  return (
+    <>
+      <div className="mb-6">
+        <h2>{t("Manage ownership")}</h2>
+
+        <p>
+          Add, remove, or transfer ownership to another GC Forms account within your department or
+          agency.
+        </p>
+
+        {message && message}
+
+        <p className="mb-4">{t("assignUsersToTemplate")}</p>
+        <p className="mb-4 font-bold">{t("enterOwnersEmail")} </p>
+        <Select
+          instanceId={useId()}
+          isClearable
+          isSearchable
+          isMulti
+          components={makeAnimated()}
+          options={usersToOptions(allUsers)}
+          value={selectedUsers}
+          onChange={(value) => setSelectedUsers(value as { value: string; label: string | null }[])}
+        />
+        <br />
+        <Button theme="secondary" type="submit" onClick={() => saveAssignations()}>
+          {t("save")}
+        </Button>
+      </div>
+    </>
+  );
+};
